requirements.txt
Flask==2.3.3
numpy>=1.24.0
Pillow>=10.0.0
Werkzeug==2.3.7
